% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_infections.R
\name{add_summary_to_infections_plot}
\alias{add_summary_to_infections_plot}
\title{FUNCTION_TITLE}
\usage{
add_summary_to_infections_plot(plot, summary, ...)
}
\arguments{
\item{plot}{DESCRIPTION.}

\item{summary}{DESCRIPTION.}

\item{...}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
