% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_dataframe.R
\name{aggregate_dataframe}
\alias{aggregate_dataframe}
\title{FUNCTION_TITLE}
\usage{
aggregate_dataframe(...)
}
\arguments{
\item{...}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
