% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_population.R
\name{allocate_results}
\alias{allocate_results}
\title{FUNCTION_TITLE}
\usage{
allocate_results(num_samples, country)
}
\arguments{
\item{num_samples}{DESCRIPTION.}

\item{country}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
