% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badtobetter-package.R
\docType{package}
\name{badtobetter-package}
\alias{badtobetter}
\alias{badtobetter-package}
\title{badtobetter: Goes from Bad to Better Code}
\description{
We demonstrate moving from bad code to better code. We include a simple single country example as well as a more complex example for multiple countries.
}
\author{
\strong{Maintainer}: Sam Abbott \email{contact@samabbott.co.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})

Authors:
\itemize{
  \item Carl Pearson \email{carl.ab.pearson@gmail.com} (\href{https://orcid.org/0000-0003-0701-7860}{ORCID})
}

}
\keyword{internal}
