% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{behavior_d}
\alias{behavior_d}
\alias{behavior_pct}
\title{Manually create a Cohen's d effect}
\usage{
behavior_d(value = 0.5, lower = NA, upper = NA)

behavior_pct(value = 1, lower = NA, upper = NA)
}
\arguments{
\item{value}{Value of the effect, either Cohen's d or percentage difference, with default}

\item{lower}{Lower boundary of 95\% confidence interval, by default NA}

\item{upper}{Upper boundary of 95\% confidence interval, by default NA}
}
\value{
Effect, class \code{better_effect}
}
\description{
Create an effect of the class \code{better_effect} based on a Cohen's d, with \code{behavior_d()}, or based on a percentage difference, with \code{behavior_pct()}.
}
\section{Functions}{
\itemize{
\item \code{behavior_pct()}: Create an effect based on a percentage difference.

}}
\examples{
behavior_d(.2)
behavior_pct(15)
}
