% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{behavior_d_to_pct}
\alias{behavior_d_to_pct}
\title{Given a behavior's Cohen's d, get its percentage difference}
\usage{
behavior_d_to_pct(effect, ..., data)
}
\arguments{
\item{effect}{A \code{better_effect} object containing the Cohen's d.}

\item{...}{Parameters to be passed on to \code{gbd_filter}.}

\item{data}{GBD dataset, of \code{gbd_dataset} class. By default, \code{gbd} is used, if it exists. To learn more, call \code{better_vignette(better)}.}
}
\value{
A \code{better_effect} object containing the percentage difference.
}
\description{
Given a behavior's Cohen's d as better_effect, get its percentage difference as better_effect. The primary use of this function is as an intermediate step between \code{nudge_to_behavior_d()} and \code{behavior_pct_to_disease()}. It is implicitly used by \code{disease()}.
}
\details{
A better_effect is a guesstimate rather than a rigorous impact analysis. Use it with caution. Check out the full calculation with \code{explanation(.)}.
The calculation uses data from the following source:
Source: Institute for Health Metrics and Evaluation. Used with permission. All rights reserved..
The authors and owners of these data retain all rights to their data. Also, they were not involved in the creation of \code{better}, and bear no responsibility for any possible issues with the representation and/or use of the data.
}
\examples{
\dontshow{
NULL
library(tidyverse)
}

nudge_to_behavior_d("information") \%>\%
  behavior_d_to_pct
}
