% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{behavior_pct_to_disease}
\alias{behavior_pct_to_disease}
\alias{disease}
\title{Given a behavior effect in percentage points, get the disease effect}
\usage{
behavior_pct_to_disease(
  effect,
  cause,
  measure = "DALYs",
  metric = "rate",
  ...,
  data
)

disease(
  effect = nudge(nudge = "all", behavior = "health"),
  cause = "all causes",
  measure = "DALYs",
  metric = "rate",
  ...
)
}
\arguments{
\item{effect}{The behavior effect to be translated into a disease effect, class \code{better_effect}}

\item{cause}{The disease, that is, the cause of mortality and/or disability}

\item{measure}{The measure for mortality/disability}

\item{metric}{The metric (rate, percentage, number) for the measure}

\item{...}{Other arguments for \code{gbd_filter}.}

\item{data}{The GDB dataset to use. By default, \code{gbd}.}
}
\value{
The effect on disease mortality and/or disability, class \code{better_effect}
}
\description{
Given a behavior effect in percentage points, get the disease effect
}
\section{Functions}{
\itemize{
\item \code{disease()}: Given a behavior effect in Cohen's d, get the disease effect

}}
\examples{
\dontshow{
options(better.gbd_path = "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com/My\ Drive/Offline\ Drive/R\ -\ Offline\ Drive/R\ packages\ JW/better/ignore/Global\ Burden\ of\ Disease\ Study\ Data")
}
\dontrun{
# If you haven't already:
# Add the following to your .Rprofile (and restart):
options(better.gbd_path = "/Users/johnwilliam/datasets/Global Burden of Disease data")
# See `vignette(\"better\")`
}

library(tidyverse)

nudge_to_behavior_d(behavior = "health", nudge = "information") \%>\%
behavior_d_to_pct \%>\%
  behavior_pct_to_disease(cause = "cardiovascular diseases")
\dontshow{
options(better.gbd_path = "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com/My\ Drive/Offline\ Drive/R\ -\ Offline\ Drive/R\ packages\ JW/better/ignore/Global\ Burden\ of\ Disease\ Study\ Data")
}
}
