% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better-package.R
\docType{package}
\name{better-package}
\alias{better}
\alias{better-package}
\title{better: Make guestimates about the impact of health nudges}
\description{
Tell `better` what kind of behavioral nudge you are considering, and what disease you are targetting, and `better` will give you guestimates about its impact on reducing deaths and disability. It primarily uses data from the Global Burden of Disease study.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/janlindemans/better}
  \item Report bugs at \url{https://github.com/janlindemans/better/issues}
}

}
\author{
\strong{Maintainer}: Jan Lindemans \email{janwillem.lindemans@gmail.com} (\href{https://orcid.org/0000-0002-3702-8651}{ORCID})

}
\keyword{internal}
