% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{better_vignette}
\alias{better_vignette}
\title{Get pre-installed \code{better} vignettes}
\usage{
better_vignette(...)
}
\arguments{
\item{...}{The name(s) of the vignette(s) you want to read, unquoted. By default, if you don't specify anything, it's all the vignettes. Options are: better, dataset-extension, gbd.}
}
\value{
Names of the vignette(s), invisible.
}
\description{
Open the HTML files of pre-installed vignettes.
}
\details{
Some \code{better} vignettes require things that you may not had set up yet at the time of installing \code{better}. Since such vignettes can't be built at the time of installation (which is how regular vignettes are created), they are pre-installed as HTML files. Use \code{better_vignette} to access them. Strictly speaking, they aren't vignettes. You can't access them with \verb{vignette(*, package = "better")}. But they are articles that fulfill the role of typical vignettes. Note that you can also use \code{better_vignette()} for true vignettes that are also accessible with \verb{vignette(*, package = "better")}. NOTE: UNDER DEVELOPMENT, ONLY ONE VIGNETTE SO FAR.
}
\examples{
\dontrun{
better_vignette()
better_vignette(better)
}
}
