% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_quantiles.R
\name{calculate_quantiles}
\alias{calculate_quantiles}
\title{FUNCTION_TITLE}
\usage{
calculate_quantiles(x)
}
\arguments{
\item{x}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
