% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{compare_effects}
\alias{compare_effects}
\title{Graph with range of effects}
\usage{
compare_effects(
  measure = get_default("measure"),
  metric = get_default("metric"),
  calculate = get_default("calculate"),
  year = get_default("year"),
  ...,
  params,
  N = 10,
  gbdeffects = gbd_effects
)
}
\arguments{
\item{measure}{Measure}

\item{metric}{Metric}

\item{calculate}{What to calculate, by default disease.}

\item{year}{Year}

\item{...}{Other}

\item{params}{Not yet developed}

\item{N}{How many comparisons, integer.}

\item{gbdeffects}{GBD effects data frame, by default \code{gbd_effects}.}
}
\value{
The data for the graph plotted, invisible.
}
\description{
Graph with range of effects
}
\examples{
compare_effects()
compare_effects(measure = "deaths")
}
