% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{cost}
\alias{cost}
\title{Get the Dollar cost of disability}
\usage{
cost(effect, ..., method = "Kennedy, Wood, & Frieden, 2017")
}
\arguments{
\item{effect}{Disease effect, of \code{better_effect} class}

\item{...}{Arguments passed on to dispatched cost function, depending on chosen method. IN DEVELOPMENT}

\item{method}{Method of cost calculation, by default \code{"Kennedy, Wood, & Frieden, 2017"}. IN DEVELOPMENT}
}
\value{
Dollar cost, of \code{better_effect} class
}
\description{
For any disease effect, get a guesstimate of its Dollar cost. STILL IN DEVELOPMENT: For now, it's a very simple guesstimate based on one general statistic. In the future, it may be smarter and take into account the different costs of different diseases.
}
\section{Source}{
The cost estimate per year of disability comes from the following paper:

Kennedy, J., Wood, E. G., & Frieden, L. (2017). Disparities in Insurance Coverage, Health Services Use, and Access Following Implementation of the Affordable Care Act: A Comparison of Disabled and Nondisabled Working-Age Adults. Inqruiry: The Journal of Health Care Organization, Provision, and Financing, 54, 004695801773403. https://doi.org/10.1177/0046958017734031
}

\examples{
\dontshow{
library(tidyverse)
NULL
}
nudge("information", behavior = "food") \%>\%
  disease("diabetes mellitus", measure = "YLDs") \%>\%
  cost
}
