% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{describe}
\alias{describe}
\title{Succinctly describe the effect}
\usage{
describe(effect)
}
\arguments{
\item{effect}{Effect to be described, \code{better_effect}}
}
\value{
The inputted effect, invisibly.
}
\description{
Succinctly describe the effect
}
\examples{
\dontshow{
library(tidyverse)
}
nudge("information") \%>\%
  describe
}
