% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{effect_only}
\alias{effect_only}
\title{Print the effect only}
\usage{
effect_only(x)
}
\arguments{
\item{x}{\code{better_effect} object to be printed.}
}
\value{
The inputted \code{better_effect} object, silently.
}
\description{
Print the effect only
}
\examples{
\dontshow{
library(tidyverse)
NULL
}
nudge("information", behavior = "food") \%>\%
  effect_only
}
