% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{etiology}
\alias{etiology}
\title{Get the etiology effect of a disease}
\usage{
etiology(effect, x, data)
}
\arguments{
\item{effect}{The disease effect (e.g., \code{"lower respiratory infections"}) to be converted into an etiology effect (e.g., \code{"influenza"}).}

\item{x}{Etiology}

\item{data}{Path of the GBD folder, where the data are stored. By default, automatically read from options, namely, \code{better.gbd_path}, ideally set in .Rprofile. To learn more, call \code{better_vignette(better)}.}
}
\value{
The etiology effect, as \code{better_effect} object
}
\description{
Convert a disease effect, like \code{"lower respiratory infections"}, into an etiology effect, like \code{"influenza"}. In our example, this boils down to investigating how influenza (the etiology) causes lower respiratory infections (the disease), which cause disability and death.
}
\examples{
\dontshow{
library(tidyverse)
NULL
}
nudge("information") \%>\%
  disease("lower respiratory infections") \%>\%
  etiology("influenza")
}
