% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{explanation}
\alias{explanation}
\title{Get explanation of \code{better_effect} object}
\usage{
explanation(x, full = TRUE, truncated = FALSE)
}
\arguments{
\item{x}{\code{better_effect} object}

\item{full}{Get full explanation or partial explanation}

\item{truncated}{Whether or not to truncate long explanations}
}
\value{
Explanation as string
}
\description{
Get explanation of \code{better_effect} object
}
\examples{
\dontshow{
options(better.gbd_path = "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com/My\ Drive/Offline\ Drive/R\ -\ Offline\ Drive/R\ packages\ JW/better/ignore/Global\ Burden\ of\ Disease\ Study\ Data")
}
\dontrun{
# If you haven't already:
# Add the following to your .Rprofile (and restart):
options(better.gbd_path = "/Users/johnwilliam/datasets/Global Burden of Disease data")
# See `vignette(\"better\")`
}

library(tidyverse)

gbd <- gbd_read()

nudge_to_behavior_d() \%>\%
  behavior_d_to_pct \%>\%
  explanation
}
