% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{facts}
\alias{facts}
\title{List the facts}
\usage{
facts(effect)
}
\arguments{
\item{effect}{The effect to be fact checked.}
}
\value{
The inputted effect, invisibly.
}
\description{
List the facts on which the guesstimate of the effect is based.
}
\examples{
\dontshow{
library(tidyverse)
}
nudge("information") \%>\%
  facts
}
