% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_behavior}
\alias{gbd_behavior}
\title{Filter Global Burden of Disease data on behavioral risks}
\usage{
gbd_behavior(..., rei = "Behavioral risks", data)
}
\arguments{
\item{...}{Parameters for \code{gbd_filter}}

\item{rei}{Risk factor for the disease, impairment, or etiology}

\item{data}{GBD dataset, of \code{gbd_dataset} class. By default, \code{gbd} is used, if it exists. To learn more, call \code{better_vignette(better)}.}
}
\value{
GBD data for given parameters, as data frame.
}
\description{
Wrapper around \code{gbd_filter}
}
\examples{
\dontshow{
NULL
}

gbd_behavior(cause = "diabetes mellitus")
}
