% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_browse}
\alias{gbd_browse}
\title{Browse the full Global Burden of Disease codebook}
\usage{
gbd_browse(data)
}
\arguments{
\item{data}{GBD dataset, of \code{gbd_dataset} class. By default, \code{gbd} is used, if it exists. To learn more, call \code{better_vignette(better)}.}
}
\value{
Last category browsed, as string.
}
\description{
Browse the GBD codebook: the available causes, or the available risk, etiologies and impairments. Since not all the GBD data is available (because it's huge), you can browse what is available on the GBD website for downloading.
}
\examples{
\dontrun{
gbd_browse()
gbd_download()
}
}
