% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_browse_cause}
\alias{gbd_browse_cause}
\title{Browse Global Burden of Disease causes}
\usage{
gbd_browse_cause(data)
}
\arguments{
\item{data}{GBD dataset, of \code{gbd_dataset} class. By default, \code{gbd} is used, if it exists. Consider \code{gbd <- gbd_read()}. To learn more, call \code{vignette("better")}.}
}
\value{
Last cause category browsed, as string.
}
\description{
Browse Global Burden of Disease causes
}
\examples{
\dontshow{
options(better.gbd_path = "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com/My\ Drive/Offline\ Drive/R\ -\ Offline\ Drive/R\ packages\ JW/better/ignore/Global\ Burden\ of\ Disease\ Study\ Data")
}
\dontrun{
# If you haven't already:
# Add the following to your .Rprofile (and restart):
options(better.gbd_path = "/Users/johnwilliam/datasets/Global Burden of Disease data")
# See `vignette(\"better\")`

# GBD functions by default read data from `gbd`. See `vignette(\"better\")`
gbd <- gbd_read()

gbd_browse_cause()
}
}
