% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_clean}
\alias{gbd_clean}
\title{Clean Global Burden of Disease data}
\usage{
gbd_clean(x)
}
\arguments{
\item{x}{Dataframe of data to be cleaned.}
}
\value{
Dataframe with cleaned data.
}
\description{
Cleans the GBD data in a sensible way.
}
\examples{
GBD_PATH <- paste0(
  "/Users/jwl38/Library/CloudStorage/Box-Box/CAH/CAH Shared/IRB Projects",
  "/Health Projects/Health Team/Public datasets on health",
  "/Global Burden of Disease Study Data"
)
gbd <- gbd_read(GBD_PATH) # better set default path in options (see ?gbd_read)
gbd <- gbd_clean(gbd)
gbd
}
