% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_codebook}
\alias{gbd_codebook}
\title{GBD codebook}
\usage{
gbd_codebook(..., data)
}
\arguments{
\item{...}{NOT YET IMPLEMENTED: Parts of the codebook you want to subset.}

\item{data}{GBD dataset as \code{gbd_dataset} object}
}
\value{
GBD codebook, list
}
\description{
GBD codebook
}
\examples{
\dontshow{
options(better.gbd_path = "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com/My\ Drive/Offline\ Drive/R\ -\ Offline\ Drive/R\ packages\ JW/better/ignore/Global\ Burden\ of\ Disease\ Study\ Data")
}
\dontrun{
# If you haven't already:
# Add the following to your .Rprofile (and restart):
options(better.gbd_path = "/Users/johnwilliam/datasets/Global Burden of Disease data")
# See `vignette(\"better\")`
}

# GBD functions by default read data from `gbd`. See `vignette(\"better\")`
gbd <- gbd_read()

gbd_codebook()
}
