% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_codebook_clean}
\alias{gbd_codebook_clean}
\title{Clean Global Burden of Disease codebook}
\usage{
gbd_codebook_clean(x = gbd_codebook_read())
}
\arguments{
\item{x}{List with codebooks to be cleaned.}
}
\value{
List with cleaned codebooks.
}
\description{
Clean Global Burden of Disease codebook
}
\examples{
\dontshow{
options(better.gbd_path = "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com/My\ Drive/Offline\ Drive/R\ -\ Offline\ Drive/R\ packages\ JW/better/ignore/Global\ Burden\ of\ Disease\ Study\ Data")
}
\dontrun{
# If you haven't already:
# Add the following to your .Rprofile (and restart):
options(better.gbd_path = "/Users/johnwilliam/datasets/Global Burden of Disease data")
# See `vignette(\"better\")`
}

library(tidyverse)

gbd_codebook <- gbd_codebook_read() \%>\%
  gbd_codebook_clean
}
