% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_codebook_read}
\alias{gbd_codebook_read}
\title{Read Global Burden of Disease codebook}
\usage{
gbd_codebook_read(path = getOption("better.gbd_path"))
}
\arguments{
\item{path}{Path of the parentfolder of the codebooks.}
}
\value{
List with dataframes of GBD codebooks.
}
\description{
Read the GBD codebook from your local folder. It looks for codebook files in all subfolders of the GBD folder and puts them into a list.
}
\examples{
# Folder where you saved your GBD codebook:
GBD_PATH <- paste0(
  "/Users/jwl38/Library/CloudStorage/Box-Box/CAH/CAH Shared/IRB Projects",
  "/Health Projects/Health Team/Public datasets on health",
  "/Global Burden of Disease Study Data"
)
gbd_meta <- gbd_codebook_read(GBD_PATH)

# Better set the default path in `options`,
# so you don't need to specify paths anymore:
options(better.gbd_path = GBD_PATH)
gbd <- gbd_codebook_read()
# Even better: specify the GBD path option in your .Rprofile file,
# so the path will be available accross sessions.
}
