% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_data}
\alias{gbd_data}
\title{Get the GBD dataset}
\usage{
gbd_data(path)
}
\arguments{
\item{path}{Path of the GBD folder, where the data are stored. By default, automatically read from options, namely, \code{better.gbd_path}, ideally set in .Rprofile. To learn more, call \code{better_vignette(better)}.}
}
\value{
GBD dataset, \code{gbd_dataset} class
}
\description{
You rarely need it. But, since there's \code{\link[=nudge_data]{nudge_data()}} and \code{\link[=nudge_codebook]{nudge_codebook()}}, there may as well be \code{\link[=gbd_data]{gbd_data()}} and \code{\link[=gbd_codebook]{gbd_codebook()}}.
}
\examples{
gbd_data()
}
