% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_data_clean}
\alias{gbd_data_clean}
\title{Clean Global Burden of Disease data}
\usage{
gbd_data_clean(x = gbd_data_read())
}
\arguments{
\item{x}{Dataframe of raw GBD data to be cleaned.}
}
\value{
Dataframe with cleaned data, by default read with \code{gbd_data_read()}.
}
\description{
Cleans the GBD data in a sensible way.
}
\examples{
\dontshow{
options(better.gbd_path = "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com/My\ Drive/Offline\ Drive/R\ -\ Offline\ Drive/R\ packages\ JW/better/ignore/Global\ Burden\ of\ Disease\ Study\ Data")
}
\dontrun{
# If you haven't already:
# Add the following to your .Rprofile (and restart):
options(better.gbd_path = "/Users/johnwilliam/datasets/Global Burden of Disease data")
# See `vignette(\"better\")`

library(tidyverse)

gbd_data <- gbd_data_read() \%>\%
  gbd_data_clean
gbd_data
}
}
