% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_data_read}
\alias{gbd_data_read}
\title{Read Global Burden of Disease data}
\usage{
gbd_data_read(path)
}
\arguments{
\item{path}{Path of the GBD folder, where the data are stored. By default, automatically read from options, namely, \code{better.gbd_path}, ideally set in .Rprofile. To learn more, call \verb{vignette(\\"better\\")}.}
}
\value{
Dataframe containing the GBD data.
}
\description{
Read the GBD data from your local folder. It looks for data files in all subfolders of the GBD folder and merges them into one dataframe.
}
\examples{
\dontshow{
options(better.gbd_path = "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com/My\ Drive/Offline\ Drive/R\ -\ Offline\ Drive/R\ packages\ JW/better/ignore/Global\ Burden\ of\ Disease\ Study\ Data")
}
\dontrun{
# If you haven't already:
# Add the following to your .Rprofile (and restart):
options(better.gbd_path = "/Users/johnwilliam/datasets/Global Burden of Disease data")
# See `vignette(\"better\")`

gbd_data <- gbd_data_read() # by default the GBD path is taken from .Rprofile
gbd_data
}
}
