% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_describe}
\alias{gbd_describe}
\title{Describe gbd_dataset object}
\usage{
gbd_describe(..., explain = TRUE, summary = FALSE, reference = FALSE)
}
\arguments{
\item{...}{Parameters determining what GBD data you want to describe. Arguments passed on to \code{gbd_filter}.}

\item{explain}{Whether or not to add an explanation of the measure, logical}

\item{summary}{Whether or not to provide a brief summary instead of a description, logical}

\item{reference}{Whether or not to add a reference, logical}
}
\value{
Description as string
}
\description{
Describe gbd_dataset object
}
\examples{
gbd_describe(cause = "diabetes mellitus")
}
