% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_download}
\alias{gbd_download}
\title{Download Global Burden of Disease data}
\usage{
gbd_download(path = getOption("better.gbd_path"))
}
\arguments{
\item{path}{Path of the parentfolder of the GBD data.}
}
\value{
URLs of webpages opened, invisible.
}
\description{
Download Global Burden of Disease data
}
\examples{
GBD_PATH <- paste0(
  "/Users/jwl38/Library/CloudStorage/Box-Box/CAH/CAH Shared/IRB Projects",
  "/Health Projects/Health Team/Public datasets on health",
  "/Global Burden of Disease Study Data"
)
gbd_download(GBD_PATH) # better set default path in options (see ?gbd_read)
}
