% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbd_effects}
\alias{gbd_effects}
\title{GBD effects dataset}
\format{
\subsection{\code{gbd_effects}}{

A data frame with 18016 rows and 15 columns:
\describe{
\item{cause}{A single disease or injury or an aggregation of diseases and injuries that causes death or disability.}
\item{rei}{Risk/etiology/impairment/injury.}
\item{metric}{The unit by which a measure is expressed. E.g., number, percent, rate, etc.}
\item{measure}{The indicator for which estimates are produced.}
\item{location}{Includes country, non-sovereign region, principal administrative unit of a country (e.g., state, province), GBD region, or other custom administrative division, such as World Bank Income Level or WHO region.}
\item{sex}{Male, female or both sexes combined.}
\item{age}{A population segment within a specified age range.}
\item{year}{The period of 365 days (or 366 days in leap years) in the Gregorian calendar divided into 12 months beginning with January and ending with December.}
\item{val}{The mean value of an estimate.}
\item{upper}{Upper value of confidence interval.}
\item{lower}{Lower value of confidence interval.}
}
}
}
\usage{
gbd_effects
}
\description{
This dataset contains estimates of effects.
}
\section{Terms of use, license and citation}{
The effects dataset is based on the GBD dataset, created and maintained by the \href{https://www.healthdata.org/}{Institute for Health Metrics and Evaluation} at the University of Washington. The dataset is owned by the University of Washington. The \code{better} package uses the dataset in accordance with the \href{https://www.healthdata.org/Data-tools-practices/data-practices/ihme-free-charge-non-commercial-user-agreement}{IHME free-of-charge non-commercial user agreement}.

Cite the original data as follows:

Global Burden of Disease Collaborative Network. Global Burden of Disease Study 2019 (GBD 2019) Results. Seattle, United States: Institute for Health Metrics and Evaluation (IHME), 2020. Available from \url{https://vizhub.healthdata.org/gbd-results/}.
}

\keyword{datasets}
