% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_filter}
\alias{gbd_filter}
\title{Subset the Global Burden of Disease dataset}
\usage{
gbd_filter(
  cause = "all causes",
  measure = "DALYs",
  metric = "rate",
  rei = NA,
  location = "Global",
  sex = "both",
  age = "all ages",
  year = max(data$year),
  data,
  params = list()
)
}
\arguments{
\item{cause}{Cause(s) of death/disability}

\item{measure}{Measure of returned values}

\item{metric}{Metric for returned values}

\item{rei}{Risk factor for the disease, impairment, or etiology}

\item{location}{Location for the data}

\item{sex}{Sex of the population}

\item{age}{Age of the population}

\item{year}{Year of data collection}

\item{data}{GBD dataset, of \code{gbd_dataset} class. By default, \code{gbd} is used, if it exists. Consider \code{gbd <- gbd_read()}. To learn more, call \code{vignette("better")}.}

\item{params}{List of parameters. Rather than feeding the parameters as arguments, like for instance gbd_filter(cause = "all causes", measure = "DALYs"), you can feed the parameters as a list, like for instance gbd_filter(params = list(cause = "all causes", measure = "DALYs"))}
}
\value{
GBD data for given parameters, as dataframe.
}
\description{
Subset the Global Burden of Disease dataset with sensible defaults. Note that \code{argument = NA} means filtering out \code{NA}s, while \code{argument = NULL} means not filtering out anything (keeping all). \code{gbd_filter} is all about the sensible defaults. If you need many arguments to be \code{NULL}, consider using \code{dplyr::filter(gbd, ...)} instead.
}
\examples{
\dontshow{
options(better.gbd_path = "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com/My\ Drive/Offline\ Drive/R\ -\ Offline\ Drive/R\ packages\ JW/better/ignore/Global\ Burden\ of\ Disease\ Study\ Data")
}
\dontrun{
# If you haven't already:
# Add the following to your .Rprofile (and restart):
options(better.gbd_path = "/Users/johnwilliam/datasets/Global Burden of Disease data")
# See `vignette(\"better\")`
}

# GBD functions by default read data from `gbd`. See `vignette(\"better\")`
gbd <- gbd_read()

gbd_filter(cause = "cardiovascular diseases")
}
