% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_license}
\alias{gbd_license}
\title{Global Burden of Disease data license}
\usage{
gbd_license(full = TRUE)
}
\arguments{
\item{full}{Whether to return full text of the license}
}
\value{
String, license
}
\description{
Global Burden of Disease data license
}
\examples{
gbd_license()
}
