% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_load}
\alias{gbd_load}
\title{Load saved Global Burden of Disease data and codebook}
\usage{
gbd_load(path = getOption("better.gbd_path"))
}
\arguments{
\item{path}{Path of the folder of the .RData file.}
}
\value{
The value of \code{base::load}, which this function wraps.
}
\description{
Load saved Global Burden of Disease data and codebook
}
\examples{
GBD_PATH <- paste0(
  "/Users/jwl38/Library/CloudStorage/Box-Box/CAH/CAH Shared/IRB Projects",
  "/Health Projects/Health Team/Public datasets on health",
  "/Global Burden of Disease Study Data"
)
gbd_load(GBD_PATH) # better set default path in options (see ?gbd_read)
}
