% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_measures}
\alias{gbd_measures}
\title{Get codebook for GBD measures}
\usage{
gbd_measures(...)
}
\arguments{
\item{...}{Which measures you want a codebook for, string.}
}
\value{
Codebook, data frame
}
\description{
Get codebook for GBD measures
}
\examples{
\dontshow{
NULL
}
gbd_measures("DALYs")
}
