% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_open}
\alias{gbd_open}
\title{Open the folder with the GBD data.}
\usage{
gbd_open(path = getOption("better.gbd_path"))
}
\arguments{
\item{path}{Path of the parentfolder of the GBD data.}
}
\value{
The value of \code{base::system2("open", path)}, which this function wraps.
}
\description{
Wrapper around \code{system2("open", path)}.
}
\examples{
GBD_PATH <- paste0(
  "/Users/jwl38/Library/CloudStorage/Box-Box/CAH/CAH Shared/IRB Projects",
  "/Health Projects/Health Team/Public datasets on health",
  "/Global Burden of Disease Study Data"
)
gbd_open(GBD_PATH) # better set default path in options (see ?gbd_read)
}
