% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_read}
\alias{gbd_read}
\title{Read cleaned Global Burden of Disease data}
\usage{
gbd_read(path, load = FALSE, assign_to)
}
\arguments{
\item{path}{Path of the GBD folder, where the data are stored. By default, automatically read from options, namely, \code{better.gbd_path}, ideally set in .Rprofile. To learn more, call \verb{vignette(\\"better\\")}.}

\item{load}{Whether or not to automatically assign the returned value to the global variable \code{gbd}. The default is \code{load = FALSE}. Rather than \code{gbd_save(load = TRUE)}, consider the cleaner alternative \code{gbd <- gbd_save()}.}

\item{assign_to}{What variable name to assign the GBD dataset to.}
}
\value{
The value of \code{base::readRDS}, which this function wraps.
}
\description{
Calls \code{base::readRDS} on the \code{.rds} file with the cleaned GBD data.
}
\examples{
\dontshow{
options(better.gbd_path = "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com/My\ Drive/Offline\ Drive/R\ -\ Offline\ Drive/R\ packages\ JW/better/ignore/Global\ Burden\ of\ Disease\ Study\ Data")
}
\dontrun{
# If you haven't already:
# Add the following to your .Rprofile (and restart):
options(better.gbd_path = "/Users/johnwilliam/datasets/Global Burden of Disease data")
# See `vignette(\"better\")`
}
gbd <- gbd_read()
}
