% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_reis}
\alias{gbd_reis}
\title{Get codebook for GBD reis}
\usage{
gbd_reis(...)
}
\arguments{
\item{...}{Which reis you want a codebook for, string.}
}
\value{
Codebook, dataframe
}
\description{
Get codebook for GBD reis
}
\examples{
\dontshow{
options(better.gbd_path = "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com/My\ Drive/Offline\ Drive/R\ -\ Offline\ Drive/R\ packages\ JW/better/ignore/Global\ Burden\ of\ Disease\ Study\ Data")
}
gbd_reis("behavioral risks")
}
