% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_save}
\alias{gbd_save}
\title{Save Global Burden of Disease data and codebook}
\usage{
gbd_save(path = getOption("better.gbd_path"))
}
\arguments{
\item{path}{Path of the parentfolder of the raw data and raw codebook, where also the .RData file with the cleaned data and codebook will be saved.}
}
\description{
Save Global Burden of Disease data and codebook
}
\examples{
GBD_PATH <- paste0(
  "/Users/jwl38/Library/CloudStorage/Box-Box/CAH/CAH Shared/IRB Projects",
  "/Health Projects/Health Team/Public datasets on health",
  "/Global Burden of Disease Study Data"
)
gbd_save(GBD_PATH) # better set default path in options (see ?gbd_read)
}
