% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_variables}
\alias{gbd_variables}
\title{Get codebook for GBD variables}
\usage{
gbd_variables(...)
}
\arguments{
\item{...}{Which variables you want a codebook for, string.}
}
\value{
Codebook, data frame
}
\description{
Get codebook for GBD variables
}
\examples{
\dontshow{
NULL
}
gbd_variables("metric")
}
