% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{net}
\alias{net}
\title{Get the net cost savings}
\usage{
net(effect, ..., method = "Benartzi, Beshears, Milkman, et al., 2017")
}
\arguments{
\item{effect}{The cost savings effect to be converted into a net cost savings effect.}

\item{...}{Currently not used, in development.}

\item{method}{The method with which to calculate the net cost.}
}
\value{
The effect on net savings, as \code{better_effect} object.
}
\description{
Calculate the net cost savings, that is, the savings in terms of reduced healthcare cost minus the costs of the nudge campaign. The \code{"Benartzi, Beshears, Milkman, et al., 2017"} method makes use of cost estimates of nudges in two categories: structural (more expensive) and nonstructural (less expensive).
}
\examples{
\dontshow{
library(tidyverse)
NULL
}
nudge("information", behavior = "food") \%>\%
  disease("diabetes mellitus", measure = "YLDs") \%>\%
  cost \%>\%
  net
}
