% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{nudge}
\alias{nudge}
\alias{nudge_to_behavior_d}
\title{Given a nudge, get its effect on behavior}
\usage{
nudge(nudge = "all", behavior = "health")

nudge_to_behavior_d(nudge = "all", behavior = "all")
}
\arguments{
\item{nudge}{Category of nudge: "information", "structure", "assistance", or "all" (any kind of nudge). By default "all". Call \code{nudge_codebook()} for more details.}

\item{behavior}{Category of behavior: "health", "food", or "all" (any kind of behavior). By default "health". Call \code{nudge_codebook()} for more details.}
}
\value{
Effect size, object of the \code{better_effect} class
}
\description{
Given a certain kind of nudge, get its effect on a certain kind of behavior
}
\details{
A better_effect is a guesstimate rather than a rigorous impact analysis. Use it with caution. Check out the full calculation with \code{explanation(.)}.
The calculation uses data from the following source:
Mertens, S., Herberz, M., Hahnel, U. J. J., & Brosch, T. (2022). The effectiveness of nudging: A meta-analysis of choice architecture interventions across behavioral domains. Proceedings of the National Academy of Sciences, 119(1), e2107346118. https://doi.org/10.1073/pnas.2107346118.
The authors and owners of these data retain all rights to their data. Also, they were not involved in the creation of \code{better}, and bear no responsibility for any possible issues with the representation and/or use of the data.
}
\section{Functions}{
\itemize{
\item \code{nudge_to_behavior_d()}: More explicit form with \code{behavior = "all"} as default

}}
\examples{
\dontshow{
NULL
library(tidyverse)
}
nudge("information")
nudge("information", behavior = "food") \%>\%
  disease("diabetes mellitus")
\dontshow{
NULL
library(tidyverse)
}

# The more explicit form allows you to write the following...
nudge(nudge = "information") \%>\%
  disease(cause = "diabetes mellitus")
# ...as this here...
nudge_to_behavior_d(nudge = "information", behavior = "health") \%>\%
  behavior_d_to_pct \%>\%
  behavior_pct_to_disease(cause = "diabetes mellitus")

}
