% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{nudge_codebook}
\alias{nudge_codebook}
\title{Get the codebook for the nudge data}
\usage{
nudge_codebook(x = quiet(nudge_data()))
}
\arguments{
\item{x}{Nudge dataset, by default automatically read.}
}
\value{
Codebook, data frame.
}
\description{
Get the codebook for the nudge data
}
\examples{
\dontshow{
NULL
}
nudge_codebook()
}
