% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{nudge_to_behavior}
\alias{nudge_to_behavior}
\title{Given a nudge, get its effect on behavior}
\usage{
nudge_to_behavior(nudge = "all", behavior = "all")
}
\arguments{
\item{nudge}{Category of nudge}

\item{behavior}{Category of behavior}
}
\value{
Effect size, object of the \code{better_class} effect
}
\description{
Given a nudge, get its effect on behavior
}
\examples{
nudge_to_behavior()
}
