% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{nudge_to_behavior_d}
\alias{nudge_to_behavior_d}
\alias{nudge}
\title{Given a nudge, get its effect on behavior}
\usage{
nudge_to_behavior_d(nudge = "all", behavior = "all")

nudge(nudge = "all", behavior = "health")
}
\arguments{
\item{nudge}{Category of nudge: "information", "structure", "assistance", or "all" (any kind of nudge)}

\item{behavior}{Category of behavior: "health", "food", or "all" (any kind of behavior)}
}
\value{
Effect size, object of the \code{better_class} effect
}
\description{
Given a certain kind of nudge, get its effect on a certain kind of behavior
}
\section{Functions}{
\itemize{
\item \code{nudge()}: Short form with \code{behavior = "health"} as default

}}
\examples{
nudge_to_behavior_d(nudge = "information", behavior = "food")
\dontshow{
options(better.gbd_path = "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com/My\ Drive/Offline\ Drive/R\ -\ Offline\ Drive/R\ packages\ JW/better/ignore/Global\ Burden\ of\ Disease\ Study\ Data")
}
nudge("information")
}
