% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{per}
\alias{per}
\title{Set the denominator for rates}
\usage{
per(effect, n = 1)
}
\arguments{
\item{effect}{Effect, \code{metric} has to be \verb{"rate}".}

\item{n}{The denominator for that rate: per how many? By default 1.}
}
\value{
Effect, \code{better_effect}.
}
\description{
Convert a rate per 100,000 into any kind of rate by specifying the denominator.
}
\examples{
\dontshow{
library(tidyverse)
NULL
}
nudge("information", behavior = "food") \%>\%
  disease("diabetes mellitus", measure = "YLDs") \%>\%
  cost \%>\%
  per(350000)
}
