% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_infections.R
\name{plot_infections}
\alias{plot_infections}
\title{FUNCTION_TITLE}
\usage{
plot_infections(pop)
}
\arguments{
\item{pop}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
