% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{effect_only}
\alias{effect_only}
\alias{print.better_effect}
\title{Print objects of \code{better_effect} class}
\usage{
effect_only(x)

\method{print}{better_effect}(x, effect_only = FALSE)
}
\arguments{
\item{x}{\code{better_effect} object}

\item{effect_only}{Whether or not to only print the effect summary. By default FALSE.}
}
\value{
Original \code{better_effect} object
}
\description{
Print objects of \code{better_effect} class
}
\section{Functions}{
\itemize{
\item \code{effect_only()}: Print effect only

}}
\examples{
\dontshow{
library(tidyverse)
options(better.gbd_path = "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com/My\ Drive/Offline\ Drive/R\ -\ Offline\ Drive/R\ packages\ JW/better/ignore/Global\ Burden\ of\ Disease\ Study\ Data")
}
nudge("information") \%>\%
  effect_only
eff <- nudge_to_behavior_d()
print(eff)
}
