% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{print.gbd_dataset}
\alias{print.gbd_dataset}
\title{Print objects of the \code{gbd_dataset} class}
\usage{
\method{print}{gbd_dataset}(x, ...)
}
\arguments{
\item{x}{\code{gbd_dataset} object}

\item{...}{Arguments passed on to \code{tibble::print.tbl_df} for the main dataframe, which is a \code{tibble}.}
}
\value{
original \code{gbd_dataset} object
}
\description{
Print objects of the \code{gbd_dataset} class
}
\examples{
\dontshow{
options(better.gbd_path = "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com/My\ Drive/Offline\ Drive/R\ -\ Offline\ Drive/R\ packages\ JW/better/ignore/Global\ Burden\ of\ Disease\ Study\ Data")
}
\dontrun{
# If you haven't already:
# Add the following to your .Rprofile (and restart):
options(better.gbd_path = "/Users/johnwilliam/datasets/Global Burden of Disease data")
# See `vignette(\"better\")`
}

gbd <- gbd_read()
print(gbd)
gbd
}
