% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{sample_size}
\alias{sample_size}
\title{Get sample size with power calculation}
\usage{
sample_size(effect, power = 0.8, alternative = "less", ...)
}
\arguments{
\item{effect}{Effect, of \code{better_effect} class.}

\item{power}{Power of the test. By default \code{.8}.}

\item{alternative}{Alternative hypothesis. By default "less". Can also be "two.sided" or "greater".}

\item{...}{Arguments passed on to test function of \code{pwr} package. Depending on whether the effect is a number or a proportion, this will be \code{pwr::pwr.t.test} or \code{pwr::pwr.2p.test}.}
}
\value{
Object of class "power.htest", from \code{pwr} package.
}
\description{
For a given effect, get the sample size through a power calculation with \code{\link[pwr:pwr.t.test]{pwr::pwr.t.test()}}, which \code{sample_size} wraps.
}
\details{
Exactly one of the parameters 'd','n','power' and
     'sig.level' must be passed as NULL, and that parameter is
     determined from the others. Notice that the last one has non-NULL
     default so NULL must be explicitly passed if you want to compute
     it.
}
\examples{
\dontshow{
library(tidyverse)
NULL
}
nudge("information", behavior = "food") \%>\%
  sample_size
}
