% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{terms_of_use}
\alias{terms_of_use}
\title{Get Terms of Use}
\usage{
terms_of_use(x, ...)
}
\arguments{
\item{x}{Object to get terms of use on}

\item{...}{Passed on to class-specific methods.}
}
\value{
Original object, invisible.
}
\description{
Get the Terms of Use for an object.
}
\examples{
terms_of_use(nudge_data())
}
