% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{tree}
\alias{tree}
\title{Create tree diagram of codebook}
\usage{
tree(x = gbd_causes())
}
\arguments{
\item{x}{Codebook. By default, gbd_causes().}
}
\value{
The tree diagram, string, invisible.
}
\description{
Create tree diagram of codebook
}
\examples{
tree()
}
