% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_infections.R
\name{ucfirst}
\alias{ucfirst}
\title{Convert the first character of a string to uppercase}
\usage{
ucfirst(str)
}
\arguments{
\item{str}{A character string.}
}
\value{
A character string with the first character converted to uppercase.
}
\description{
This function converts the first character of a string to
uppercase.
}
\examples{
badtobetter:::ucfirst("hello")
}
