% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_dataframe.R
\name{aggregate_dataframe}
\alias{aggregate_dataframe}
\title{Convert the results of multi-column (output) \code{\link[=aggregate]{aggregate()}} to pure data.frame}
\usage{
aggregate_dataframe(...)
}
\arguments{
\item{...}{Pass the same arguments as you would to \code{\link[=aggregate]{aggregate()}}.}
}
\value{
A data.frame with the same columns as the output of \code{\link[=aggregate]{aggregate()}}.
}
\description{
This function converts the results of multi-column (output)
\code{\link[=aggregate]{aggregate()}} to pure data.frame.
}
\examples{
aggregate_dataframe(mpg ~ cyl, mtcars, mean)
}
