% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{behavior_d}
\alias{behavior_d}
\alias{behavior_pct}
\title{Create an effect based on a Cohen's d}
\usage{
behavior_d(value = 0.5, lower = NA, upper = NA)

behavior_pct(value, lower = NA, upper = NA)
}
\arguments{
\item{value}{Value of the effect}

\item{lower}{Lower boundary of 95\% confidence interval}

\item{upper}{Upper boundary of 95\% confidence interval}
}
\value{
Effect
}
\description{
Create an effect based on a Cohen's d
}
\section{Functions}{
\itemize{
\item \code{behavior_pct()}: Create an effect based on a percentage decrease

}}
\examples{
behavior_d(.2)
behavior_pct(15)
}
