% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{behavior_d_to_pct}
\alias{behavior_d_to_pct}
\title{Given a behavior's Cohen's d, get its percentage difference}
\usage{
behavior_d_to_pct(effect, ..., data)
}
\arguments{
\item{effect}{A \code{better_effect} object containing the Cohen's d.}

\item{...}{Parameters to be passed on to \code{gbd_filter}.}

\item{data}{GBD data as \code{gbd_dataset} object.}
}
\value{
A \code{better_effect} object containing the percentage difference.
}
\description{
Given a behavior's Cohen's d, get its percentage difference
}
\examples{
\dontshow{
options(better.gbd_path = "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com/My\ Drive/Offline\ Drive/R\ -\ Offline\ Drive/R\ packages\ JW/better/ignore/Global\ Burden\ of\ Disease\ Study\ Data")
}
\dontrun{
# If you haven't already:
# Add the following to your .Rprofile (and restart):
options(better.gbd_path = "/Users/johnwilliam/datasets/Global Burden of Disease data")
# See `vignette(\"better\")`
}

library(tidyverse)

# GBD functions by default read data from `gbd`. See `vignette(\"better\")`
gbd <- gbd_read()

nudge_to_behavior_d() \%>\%
  behavior_d_to_pct
}
