% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_quantiles.R
\name{calculate_quantiles}
\alias{calculate_quantiles}
\title{Calculate pre-specified quantiles}
\usage{
calculate_quantiles(x, ...)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A named numeric vector with the following names:
\itemize{
\item \code{lo.q}: The lower quartile.
\item \code{med}: The median.
\item \code{hi.q}: The upper quartile.
}
}
\description{
This function calculates pre-specified quantiles
using the \code{\link[stats:quantile]{stats::quantile()}} function.
}
\examples{
calculate_quantiles(c(1, 2, 3, 4, 5))
}
