% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{explanation}
\alias{explanation}
\title{Get explanation of \code{better_effect} object}
\usage{
explanation(x, full = TRUE, truncated = FALSE)
}
\arguments{
\item{x}{\code{better_effect} object}

\item{full}{Get full explanation of all steps or only last step}

\item{truncated}{Whether or not to truncate long explanations}
}
\value{
Explanation as string
}
\description{
Get explanation of \code{better_effect} object
}
\examples{
\dontshow{
NULL
library(tidyverse)
}

nudge_to_behavior_d(nudge = "information", behavior = "health") \%>\%
  behavior_d_to_pct \%>\%
  explanation
}
