% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_causes}
\alias{gbd_causes}
\title{Get codebook for GBD causes}
\usage{
gbd_causes(...)
}
\arguments{
\item{...}{Which causes you want a codebook for, string.}
}
\value{
Codebook, data frame
}
\description{
Get codebook for GBD causes
}
\examples{
\dontshow{
NULL
}
gbd_causes("diabetes mellitus")
}
