% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_codebook}
\alias{gbd_codebook}
\title{GBD codebook}
\usage{
gbd_codebook(..., data)
}
\arguments{
\item{...}{NOT YET IMPLEMENTED: Parts of the codebook you want to subset.}

\item{data}{GBD dataset, of \code{gbd_dataset} class. By default, \code{gbd} is used, if it exists. To learn more, call \code{better_vignette(better)}.}
}
\value{
GBD codebook, list
}
\description{
GBD codebook
}
\examples{
\dontshow{
NULL
}

gbd_codebook()
}
