% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_codebook_clean}
\alias{gbd_codebook_clean}
\title{Clean Global Burden of Disease codebook}
\usage{
gbd_codebook_clean(x)
}
\arguments{
\item{x}{List with codebooks to be cleaned.}
}
\value{
List with cleaned codebooks.
}
\description{
Clean Global Burden of Disease codebook
}
\examples{
GBD_PATH <- paste0(
  "/Users/jwl38/Library/CloudStorage/Box-Box/CAH/CAH Shared/IRB Projects",
  "/Health Projects/Health Team/Public datasets on health",
  "/Global Burden of Disease Study Data"
)
gbd_meta <- gbd_codebook_read(GBD_PATH) # better set default path in options (see ?gbd_read)
gbd_codebook_clean(gbd_meta)
}
