% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_codebook_read}
\alias{gbd_codebook_read}
\title{Read Global Burden of Disease codebook}
\usage{
gbd_codebook_read(path)
}
\arguments{
\item{path}{Path of the GBD folder, where the data are stored. By default, automatically read from options, namely, \code{better.gbd_path}, ideally set in .Rprofile. To learn more, call \verb{vignette(\\"better\\")}.}
}
\value{
List with dataframes of GBD codebooks.
}
\description{
Read the GBD codebook from your local folder. It looks for codebook files in all subfolders of the GBD folder and puts them into a list.
}
\examples{
\dontshow{
options(better.gbd_path = "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com/My\ Drive/Offline\ Drive/R\ -\ Offline\ Drive/R\ packages\ JW/better/ignore/Global\ Burden\ of\ Disease\ Study\ Data")
}
\dontrun{
# If you haven't already:
# Add the following to your .Rprofile (and restart):
options(better.gbd_path = "/Users/johnwilliam/datasets/Global Burden of Disease data")
# See `vignette(\"better\")`
}

gbd_codebook <- gbd_codebook_read()
gbd_codebook
}
