% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_download}
\alias{gbd_download}
\title{Download Global Burden of Disease data}
\usage{
gbd_download(path)
}
\arguments{
\item{path}{Path of the GBD folder, where the data are stored. By default, automatically read from options, namely, \code{better.gbd_path}, ideally set in .Rprofile. To learn more, call \code{better_vignette(better)}.}
}
\value{
URLs of webpages opened, invisible.
}
\description{
Download Global Burden of Disease data
}
\examples{
\dontshow{
NULL
}

# gbd_download()
}
