% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_filter}
\alias{gbd_filter}
\title{Subset the Global Burden of Disease dataset}
\usage{
gbd_filter(
  cause = get_default("cause"),
  measure = get_default("measure"),
  metric = get_default("metric"),
  rei = get_default("rei"),
  location = get_default("location"),
  sex = get_default("sex"),
  age = get_default("age"),
  year = get_default("year"),
  data = NA,
  params = list()
)
}
\arguments{
\item{cause}{Cause(s) of death/disability}

\item{measure}{Measure of returned values}

\item{metric}{Metric for returned values}

\item{rei}{Risk factor for the disease, impairment, or etiology}

\item{location}{Location for the data}

\item{sex}{Sex of the population}

\item{age}{Age of the population}

\item{year}{Year of data collection}

\item{data}{GBD dataset, of \code{gbd_dataset} class. By default, \code{gbd} is used, if it exists. To learn more, call \code{better_vignette(better)}.}

\item{params}{List of parameters. Rather than feeding the parameters as arguments, like for instance gbd_filter(cause = "all causes", measure = "DALYs"), you can feed the parameters as a list, like for instance gbd_filter(params = list(cause = "all causes", measure = "DALYs"))}
}
\value{
GBD data for given parameters, as data frame.
}
\description{
Subset the Global Burden of Disease dataset with sensible defaults. Note that \code{argument = NA} means filtering out \code{NA}s, while \code{argument = NULL} means not filtering out anything (keeping all). \code{gbd_filter} is all about the sensible defaults. If you need many arguments to be \code{NULL}, consider using \code{dplyr::filter(gbd, ...)} instead.
}
\section{Terms of use}{
The GBD dataset was created and is maintained by the \href{https://www.healthdata.org/}{Institute for Health Metrics and Evaluation} at the University of Washington. The dataset is owned by the University of Washington, which reserves all rights to it.

Cite the data as follows:

Global Burden of Disease Collaborative Network. \emph{Global Burden of Disease Study 2019 (GBD 2019) Results.} Seattle, United States: Institute for Health Metrics and Evaluation (IHME), 2020. Available from \url{https://vizhub.healthdata.org/gbd-results/}.

The \code{better} package uses the dataset in accordance with the \href{https://www.healthdata.org/Data-tools-practices/data-practices/ihme-free-charge-non-commercial-user-agreement}{IHME free-of-charge non-commercial user agreement}. The authors and owners of the GBD data were not involved in the creation of \code{better}, and bear no responsibility for any possible issues with the representation and/or use of the data.
}

\examples{
\dontshow{
NULL
}

gbd_filter(cause = "diabetes mellitus")
}
