% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_load}
\alias{gbd_load}
\title{Load cleaned Global Burden of Disease data}
\usage{
gbd_load(..., assign_to = "gbd")
}
\arguments{
\item{...}{Passed on to \code{gbd_read}.}

\item{assign_to}{What variable name to assign the GBD dataset to.}
}
\value{
The value of \code{base::readRDS}, which this function wraps.
}
\description{
Load cleaned Global Burden of Disease data
}
\examples{
\dontshow{
options(better.gbd_path = "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com/My\ Drive/Offline\ Drive/R\ -\ Offline\ Drive/R\ packages\ JW/better/ignore/Global\ Burden\ of\ Disease\ Study\ Data")
}
\dontrun{
# If you haven't already:
# Add the following to your .Rprofile (and restart):
options(better.gbd_path = "/Users/johnwilliam/datasets/Global Burden of Disease data")
# See `vignette(\"better\")`
}

gbd_load()
}
