% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_read}
\alias{gbd_read}
\title{Read Global Burden of Disease data}
\usage{
gbd_read(path = as.character(getOption("better.gbd_path")))
}
\arguments{
\item{path}{Path of the parentfolder of the data.}
}
\value{
Dataframe containing the GBD data.
}
\description{
Read the GBD data from your local folder. It looks for data files in all subfolders of the GBD folder and merges them into one dataframe.
}
\examples{
# Folder where you saved your GBD data:
GBD_PATH <- paste0(
  "/Users/jwl38/Library/CloudStorage/Box-Box/CAH/CAH Shared/IRB Projects",
  "/Health Projects/Health Team/Public datasets on health",
  "/Global Burden of Disease Study Data"
)
gbd <- gbd_read(GBD_PATH)
gbd

# Better set the default path in `options`,
# so you don't need to specify paths anymore:
options(better.gbd_path = GBD_PATH)
gbd <- gbd_read()
# Even better: specify the GBD path option in your .Rprofile file,
# so the path will be available accross sessions.

gbd_meta <- gbd_codebook_read()
gbd_meta
gbd_meta <- gbd_codebook_clean(gbd_meta)
gbd_meta
attach(gbd_meta)
gbd_cb
gbd_cb_cause

}
