% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_reis}
\alias{gbd_reis}
\title{Get codebook for GBD reis}
\usage{
gbd_reis(...)
}
\arguments{
\item{...}{Which reis you want a codebook for, string.}
}
\value{
Codebook, data frame
}
\description{
Get codebook for GBD reis
}
\examples{
\dontshow{
NULL
}
gbd_reis("behavioral risks")
}
