% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_save}
\alias{gbd_save}
\title{Save Global Burden of Disease data and codebook}
\usage{
gbd_save(path)
}
\arguments{
\item{path}{Path of the GBD folder, where the data are stored. By default, automatically read from options, namely, \code{better.gbd_path}, ideally set in .Rprofile. To learn more, call \verb{vignette(\\"better\\")}.}
}
\description{
Save Global Burden of Disease data and codebook
}
\examples{
\dontshow{
options(better.gbd_path = "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com/My\ Drive/Offline\ Drive/R\ -\ Offline\ Drive/R\ packages\ JW/better/ignore/Global\ Burden\ of\ Disease\ Study\ Data")
}
\dontrun{
# If you haven't already:
# Add the following to your .Rprofile (and restart):
options(better.gbd_path = "/Users/johnwilliam/datasets/Global Burden of Disease data")
# See `vignette(\"better\")`

gbd_save()
}
}
