% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{gbd_variables}
\alias{gbd_variables}
\title{Get codebook for GBD variables}
\usage{
gbd_variables(...)
}
\arguments{
\item{...}{Which variables you want a codebook for, string.}
}
\value{
Codebook, dataframe
}
\description{
Get codebook for GBD variables
}
\examples{
\dontshow{
options(better.gbd_path = "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com/My\ Drive/Offline\ Drive/R\ -\ Offline\ Drive/R\ packages\ JW/better/ignore/Global\ Burden\ of\ Disease\ Study\ Data")
}
gbd_variables("metric")
}
