% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{nudge_codebook}
\alias{nudge_codebook}
\title{Get the codebook for the nudge data}
\usage{
nudge_codebook(x = nudge_data())
}
\arguments{
\item{x}{Nudge dataset, by default automatically read.}
}
\value{
Codebook, dataframe.
}
\description{
Get the codebook for the nudge data
}
\examples{
\dontshow{
options(better.gbd_path = "/Users/jwl38/Library/CloudStorage/GoogleDrive-janwillem.lindemans@gmail.com/My\ Drive/Offline\ Drive/R\ -\ Offline\ Drive/R\ packages\ JW/better/ignore/Global\ Burden\ of\ Disease\ Study\ Data")
}
nudge_codebook()
}
