% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{nudge_data}
\alias{nudge_data}
\title{Get data on the effect sizes of nudges}
\usage{
nudge_data()
}
\value{
Data on the effect sizes of nudges
}
\description{
This data is used by other \code{better} functions.
}
\examples{
nudge_data()
}
