% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbd_clean.R
\name{print.better_effect}
\alias{print.better_effect}
\title{Print objects of \code{better_effect} class}
\usage{
\method{print}{better_effect}(x)
}
\arguments{
\item{x}{\code{better_effect} object}
}
\value{
Original \code{better_effect} object
}
\description{
Print objects of \code{better_effect} class
}
\examples{
eff <- nudge_to_behavior()
print(eff)
}
