% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize_population.R
\name{synthesize_population}
\alias{synthesize_population}
\title{FUNCTION_TITLE}
\usage{
synthesize_population(
  num_samples = 1000,
  prob_high_risk,
  prob_exposure_hr,
  prob_exposure_lr,
  max_exposure_years = 30,
  num_strains = 4,
  country
)
}
\arguments{
\item{num_samples}{DESCRIPTION.}

\item{prob_high_risk}{DESCRIPTION.}

\item{prob_exposure_hr}{DESCRIPTION.}

\item{prob_exposure_lr}{DESCRIPTION.}

\item{max_exposure_years}{DESCRIPTION.}

\item{num_strains}{DESCRIPTION.}

\item{country}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
