% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_infections.R
\name{ucfirst}
\alias{ucfirst}
\title{FUNCTION_TITLE}
\usage{
ucfirst(str)
}
\arguments{
\item{str}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
